* removed MobilizedBody references in favor of new handle in ConditionalConstraint.h.  * Made SemiExplicitEulerTimeStepper API more like the Integrator (should derive from the Integrator).  * Update RigidContact1.cpp example to match API change above.

+++ Simbody/src/SemiExplicitEulerTimeStepper.cpp
@@ -412,7 +412,7 @@ stepTo(Real time) {
     cout << "  constraint status:\n";
     for (unsigned i=0; i < m_uniContact.size(); ++i) {
         ImpulseSolver::UniContactRT& rt = m_uniContact[i];
-        printf("  %d: cont %s fric %s\n", rt.m_ucx,
+        printf("  %d: cont %s fric %s\n", (int)rt.m_ucx,
                ImpulseSolver::getUniCondName(rt.m_contactCond),
                ImpulseSolver::getFricCondName(rt.m_frictionCond));
     }
